add clean feature adversaries

+++ tda/precision.py
@@ -1,3 +1,4 @@
 import torch
 
-default_tensor_type = torch.FloatTensor
+default_tensor_type = torch.DoubleTensor
+#default_tensor_type = torch.FloatTensor
